+++ pychron/processing/permutator/permutator.py
@@ -1,14 +1,15 @@
-import os
 import itertools
 
 from traits.api import Property, cached_property
-from uncertainties import ufloat, nominal_value, std_dev
+
+from uncertainties import nominal_value, std_dev
 import yaml
-from numpy import array
+
 from pychron.core.helpers.logger_setup import logging_setup
+from pychron.core.progress import progress_loader
 from pychron.core.stats.core import calculate_weighted_mean
 from pychron.loggable import Loggable
-from pychron.processing.arar_age import ArArAge
+from pychron.processing.analyses.file_analysis import FileAnalysis
 
 
 # class PermutationResults(object):
@@ -23,30 +24,67 @@ from pychron.processing.arar_age import ArArAge
 #     # def __init__(self, a, *args, **kw):
 #     #     super(PermutatedAnalysis, self).__init__(*args, **kw)
 #     #     self.isotopes =
+from pychron.processing.tasks.figures.editors.ideogram_editor import IdeogramEditor
+from pychron.pychron_constants import ARGON_KEYS
+
+
+class PermutationRecord(object):
+    __slots__ = ('age', 'info_str')
 
 
 class FitPermutator(Loggable):
-    permutations = Property
-
-    def gen_permutations(self):
-        n = 5
-        ifits = [self.fits for i in range(n)]
-        return itertools.product(*ifits)
 
     def permutate(self, ai):
-        f = self._permutate
-        ages = [f(ai, perm) for perm in self.gen_permutations()]
-        xs, es = zip(*((nominal_value(x), std_dev(x)) for x in ages))
-        wm, we = calculate_weighted_mean(xs, es)
-        return wm, ages
+        func = lambda x, prog, i, n: self._permutate(ai, x, prog, i, n)
 
-    def _permutate(self, ai, perm):
-        for iso, p in zip(ai.isotopes.itervalues(), perm):
+        perms = self._gen_unique_permutations(ai.isotopes)
+        records = progress_loader(perms, func)
+
+        xs, es = zip(*((nominal_value(r.age), std_dev(r.age)) for r in records))
+        wm, we = calculate_weighted_mean(xs, es)
+        return wm, records
+
+    def _gen_unique_permutations(self, isos):
+        n = 5
+
+        skips = self.fits['skips']
+        ifits = (self.fits['fits'] for _ in range(n))
+
+        rs = []
+        for perm in itertools.product(*ifits):
+            nperm = []
+            for k, p in zip(ARGON_KEYS, perm):
+                iso = isos[k]
+                if k in skips:
+                    p = iso.fit
+                nperm.append((k, p))
+            if nperm not in rs:
+                rs.append(nperm)
+
+        return rs
+
+    def _permutate(self, ai, perm, prog, i, n):
+        isos = ai.isotopes
+        ps = []
+        for k, p in perm:
+            iso = isos[k]
             iso.set_fit(p)
             iso.dirty = True
+            ps.append(p[0])
+
         ai.calculate_age(force=True)
-        self.debug('age: {:<20s} permutation: {}'.format(str(ai.uage), ','.join(perm)))
-        return ai.uage
+
+        permstr = ','.join(ps)
+        agestr = str(ai.uage)
+        record_id = ai.record_id
+        self.debug('{} age: {:<20s} permutation: {}'.format(record_id, agestr, permstr))
+        if prog:
+            prog.change_message('Permutated {}: age: {}, perm:{}'.format(record_id,
+                                                                         agestr, permstr))
+        r = PermutationRecord()
+        r.age = ai.uage
+        r.info_str = '{} ({})'.format(record_id, permstr)
+        return r
 
 
 class Permutator(Loggable):
@@ -61,14 +99,30 @@ class Permutator(Loggable):
             self.warning('Invalid configuration file {}. error: {}'.format(self.path, e))
 
     def get_fits(self):
-        return self.configuration_dict.get('permutations').get('fits')
+        return self.configuration_dict.get('permutations').get('fit')
 
     def fits_permutation(self):
         fp = FitPermutator()
         fp.fits = self.get_fits()
         for ai in self.oanalyses:
-            wm, ages = fp.permutate(ai)
-            print wm
+            wm, records = fp.permutate(ai)
+            ans = self._make_analyses(records)
+
+            editor = IdeogramEditor()
+            po = editor.plotter_options_manager.plotter_options
+            editor.analyses = ans
+            editor.disable_aux_plots()
+            po.set_aux_plot_height('Analysis Number Stacked', 300)
+
+            editor.rebuild()
+            editor.configure_traits()
+
+    def _make_analyses(self, records):
+        return [FileAnalysis(age=nominal_value(ai.age),
+                             age_err=std_dev(ai.age),
+                             record_id=ai.info_str)
+                for ai in records]
+
 
 if __name__ == '__main__':
     p = Permutator()
@@ -78,20 +132,25 @@ if __name__ == '__main__':
     from pychron.database.isotope_database_manager import IsotopeDatabaseManager
 
     class Record(object):
-        uuid = 'd48e3d9f-4d46-4af6-8815-4820a8e62a1b'
         analysis_type = 'unknown'
+
+        def __init__(self, u):
+            self.uuid = u
 
     man = IsotopeDatabaseManager(bind=False, connect=False)
     db = man.db
     db.trait_set(kind='mysql',
-                 host='129.138.12.160',
-                 name='pychrondata',
-                 user='root', password='DBArgon')
+                 host='localhost',
+                 name='pychrondata_dev',
+                 username='root', password='Argon')
     db.connect()
 
-    a = man.make_analysis(Record(), use_cache=False, unpack=True)
-    a.j = ufloat(1e-4, 1e-7)
-    p.oanalyses = [a]
+    ans = man.make_analyses([Record('65c1c4a9-e317-452b-9654-3f06efcbe664'),
+                             Record('39b6e623-e178-4dc4-bf5c-14c81485bd54')],
+                            use_cache=False, unpack=True)
+    # a.j = ufloat(1e-4, 1e-7)
+
+    p.oanalyses = ans[:1]
 
     p.fits_permutation()